sistemInventori/requirements.txt
asgiref
Django==4.2.7
django-filter
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.1
Pillow==10.1.0
PyJWT==2.8.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
